requirements.txt
pyserial==3.5
plyer==2.1.0
streamlit==1.29.0
jinja2==3.1.2
psutil==5.9.5